pin: omit board/chip namespace when parsing and saving the pin name (#55)
board: and chip: can be prefixed to the pin to force using a pin defined in the
board or in the chip. Currently though, neither chips or boards are handling
this prefix.
Since this is just used to give priority when parsing the pin, this commit
removes the namespace from the spec passed to parsePin. This isolates the
namespace handling to the Pin type creation, without having to handle it in
chips and boards.