Use root build for chips, hals, and boards
The reason behind this change is that I'm unable to specify a chip, hal,
or board definition in a local project and that's a pattern that we
explicitly want to handle. This means packages must provide an absolute
path to their files, which is already done throughout our small
ecosystem.

In the future I'd like to change the API so package creators or those
vendoring everything locally give a relative path.