usb: unify handling of length parameters (#177)
* usb: unify handling of length parameters

The length parameters in USB packets are strict and the existing
implementation does already hardcode the length within the serialize
functions. Therefore the user should not provide these lengths during
creation time. In order to make this more verbose use defaults values
within the structs and reference these defaults values within the
serialization functions. It's not possible to remove these fields
completly as this fields are used within deserialization from the
hardware itself.

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>

* Remove length fields as they are obsolete. The lengths is only required
during the serialization and therefore is already hardcoded. The only
other convinience is the calculation of the descriptor size, but this is
already wrong. extern structs are padded and therefore do not have the
right sizes.

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>

* cleanup some accidentially added changes

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>

---------

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>