Don't distinguish between Regz generated code for "standalone"
Fixes #740.

We had a flag to make Regz generate code "for MicroZig". The point was
to be able to output code that isn't tightly coupled to MicroZig,
however the coupling is quite weak.

Instead I've removed that edge case so the imports for generated code
always looks the same. To help those trying to use Regz output for a
standalone project, I wrote a quick guide on what to implement for the
`microzig` import.