Set register and bitfield defaults to reset values
Fixes #571

This patch takes the reset value and mask found in the Regz database and
uses them for default values in the generated code. It doesn't cover
every edge case, but it's not important or useful in some of those edge
cases, such as arrays of integer registers. The stated goal of #571 is
to make invocations of the MMIO `write()` method more sane.

It also starts fixing some of the generation tests so that they work
with the writergate, and virtual filesystem changes.