Generate root from start.zig
Fixes #465.

One of the weaknesses of MicroZig's build system is that packages that
use root declarations for comptime-configuration don't work. We make the
application a package to the real root of an executable, and break out
useful stdlib options (namely logging) through `microzig_options`. This
is also notable because MicroZig changes the defaults for logging
because logging will cause compiler errors from not having an operating
system.

I want to avoid boiler-plate like the following:

```zig
const microzig = @import("microzig");

comptime {
    microzig.cpu.export_startup_logic(); // would export startup logic
}

pub fn main() {
    // ...
}
```

So this patch creates a code generation step in our build system that
takes the existing `startup.zig` code that currently makes up the root
of the executable and appends the public declarations from the
application. This gives us everything we want, no boiler plate, and
packages will now be able to use comptime-configuration.

The special handling of the Standard Library logging function is
temporary. This needs to be fixed upstream so that the regular logging
function is a no-op on freestanding targets.