Patch file reorganization
We will now use ZON patch files instead of directly importing them into
build scripts, serializing to ndjson, and sending them to Regz. Regz
will allow for multiple patch files to be set, which are applied in
order. This allows for patterns like with RP2350, where the `hazard3`
target needs to append a single patch on top of the regulare RP2350
patches which changes the processor.

ZON is easier to hand write than JSON, and possible to leave comments. I
could go for something a bit less verbose, but if we ever decided to
change, it would be easy to make a conversion script. It also has great
first class parsing/serializing support in Zig.

This organization of patches is where I want things before I start to
build tooling for it in `sorcerer`. The build system currently provides
`sorcerer` with all the locations of register schemas on the filesystem, even
the ones pulled in through dependencies. We'll now be able to get the
list of patches for a given target as well.